% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_affine_transform.R
\name{apply_affine_transform}
\alias{apply_affine_transform}
\title{Apply Affine Transformation to Points}
\usage{
apply_affine_transform(points, transform_matrix)
}
\arguments{
\item{points}{A data frame or matrix with two columns representing the coordinates of the points (X, Y).}

\item{transform_matrix}{A 3x3 matrix representing the affine transformation matrix.}
}
\value{
A data frame with the transformed coordinates (X, Y).
}
\description{
This function applies an affine transformation to a set of points using a given transformation matrix.
}
\examples{
points <- data.frame(X = c(1, 2, 3), Y = c(4, 5, 6))
transform_matrix <- matrix(c(1, 0, 0, 0, 1, 0, 1, 1, 1), nrow = 3, byrow = TRUE)
apply_affine_transform(points, transform_matrix)
}
