% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_affine_transform.R
\name{calculate_affine_transform}
\alias{calculate_affine_transform}
\title{Calculate Affine Transformation Matrix}
\usage{
calculate_affine_transform(src, dst)
}
\arguments{
\item{src}{A matrix or data frame with three rows and two columns representing the coordinates of the source points (X, Y).}

\item{dst}{A matrix or data frame with three rows and two columns representing the coordinates of the destination points (X, Y).}
}
\value{
A 3x3 matrix representing the affine transformation matrix.
}
\description{
This function calculates the affine transformation matrix that maps a set of source points to a set of destination points.
}
\examples{

src <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 3, byrow = TRUE)
dst <- matrix(c(7, 8, 9, 10, 11, 12), nrow = 3, byrow = TRUE)
calculate_affine_transform(src, dst)
}
