% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeference_image.R
\name{georeference_image}
\alias{georeference_image}
\title{Georeference an Image}
\usage{
georeference_image(ref_points_image, ref_points_world)
}
\arguments{
\item{ref_points_image}{A matrix of reference points in the image. Each row should represent a point (x, y).}

\item{ref_points_world}{A matrix of reference points in the real world. Each row should represent a point (x, y).}
}
\value{
A list containing:
\describe{
\item{transform_matrix}{The affine transformation matrix.}
\item{scale_x}{The scale factor in the x direction.}
\item{scale_y}{The scale factor in the y direction.}
\item{scale}{The average scale factor.}
\item{angle}{The rotation angle in degrees.}
}
}
\description{
This function calculates the affine transformation matrix needed to georeference an image
using provided reference points in the image and their corresponding points in the real world.
}
\examples{
# Example usage:
ref_points_image <- matrix(c(10, 20, 30, 40, 50, 60), ncol = 2, byrow = TRUE)
ref_points_world <- matrix(c(100, 200, 300, 400, 500, 600), ncol = 2, byrow = TRUE)
result <- georeference_image(ref_points_image, ref_points_world)
print(result)

}
