% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Calculate and Plot Metrics from a Data Frame}
\usage{
metrics(file, scale = 1, angle_cor = 0)
}
\arguments{
\item{file}{A data frame containing the data to be processed.}

\item{scale}{a scale factor from georeference_image function.}

\item{angle_cor}{Angle of the object to correct the directions}
}
\value{
A ggplot object with histograms of the calculated metrics.
}
\description{
This function calculates additional metrics from a data frame, filters the data, and creates histograms of the metrics.
}
\examples{
df <- data.frame(sample = c('sample_1', 'sample_2'),Name=c("A","B"), radius_max = c(10, 20), radius_min = c(2, 5),angle = c(0.5, -0.5), perimeter=c(10,2),area = c(500, 10000),eccentricity=c(2,1.3),minor_axis=c(0.1,0.2),center_z=c(0,0), center_x=c(1000,1000), center_y=c(0,0),id=c(1,2))
metrics(df)
}
