% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_image.R
\name{process_image}
\alias{process_image}
\title{Process an Image to Extract and Visualize Object Centroids}
\usage{
process_image(image_path, name = "sample_")
}
\arguments{
\item{image_path}{A string specifying the path to the image file.}

\item{name}{A string specifying the prefix for naming samples. Default is 'sample_'.}
}
\value{
A list containing the plot with centroids marked, the image width and height, and a data frame of valid centroids.
}
\description{
This function processes an image to extract the centroids of objects based on their blue channel,
filters the objects based on specific criteria, and generates a plot with the centroids marked.
}
\examples{
img_path <- system.file("extdata", "zircon.png", package = "DZubc")
result <- process_image(img_path)

print(result$plot)
}
