% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_affine_transform.R
\name{apply_affine_transform}
\alias{apply_affine_transform}
\title{Apply Affine Transformation to Points}
\usage{
apply_affine_transform(points, transform_matrix)
}
\arguments{
\item{points}{A data frame or matrix with two columns representing the coordinates of the points (X, Y).}

\item{transform_matrix}{A 3x3 matrix representing the affine transformation matrix.}
}
\value{
A data frame with the transformed coordinates (X, Y).
}
\description{
This function applies an affine transformation to a set of points using a given transformation matrix.
}
\examples{
ICP_coordinates<- matrix(c(16.1990,22.884,14.9028,84.7310,79.9283,80.9070),nrow=3)

IMAGE_coordinates<- matrix(c(2085.685,1375.439,2909.453,2846.163,1508.357,2023.621),nrow=3)

transform_matrix <- georeference_image(ICP_coordinates, IMAGE_coordinates)

FULL_IMAGE_COORDINATES<-matrix(c(seq(1500,2500,by=50),Y=seq(1750,2750,by=50)),nrow=21)

apply_affine_transform(FULL_IMAGE_COORDINATES, transform_matrix$transform_matrix)
}
