% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_affine_transform.R
\name{calculate_affine_transform}
\alias{calculate_affine_transform}
\title{Calculate Affine Transformation Matrix}
\usage{
calculate_affine_transform(src, dst)
}
\arguments{
\item{src}{A matrix or data frame with three rows and two columns representing the coordinates of the source points (X, Y).}

\item{dst}{A matrix or data frame with three rows and two columns representing the coordinates of the destination points (X, Y).}
}
\value{
A 3x3 matrix representing the affine transformation matrix.
}
\description{
This function calculates the affine transformation matrix that maps a set of source points to a set of destination points.
}
\examples{

src<- matrix(c(16.1990,22.884,14.9028,84.7310,79.9283,80.9070),nrow=3)

dst<- matrix(c(2085.685,1375.439,2909.453,2846.163,1508.357,2023.621),nrow=3)

calculate_affine_transform(src, dst)
}
