% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeference_image.R
\name{georeference_image}
\alias{georeference_image}
\title{Georeference an Image}
\usage{
georeference_image(ref_points_image, ref_points_world)
}
\arguments{
\item{ref_points_image}{A matrix of reference points in the image. Each row should represent a point (x, y).}

\item{ref_points_world}{A matrix of reference points in the real world. Each row should represent a point (x, y).}
}
\value{
A list containing:
\describe{
\item{transform_matrix}{The affine transformation matrix.}
\item{scale_x}{The scale factor in the x direction.}
\item{scale_y}{The scale factor in the y direction.}
\item{scale}{The average scale factor.}
\item{angle}{The rotation angle in degrees.}
}
}
\description{
This function calculates the affine transformation matrix needed to georeference an image
using provided reference points in the image and their corresponding points in the real world.
}
\examples{
# Example usage:
ICP_coordinates<- matrix(c(16.1990,22.884,14.9028,84.7310,79.9283,80.9070),nrow=3)

IMAGE_coordinates<- matrix(c(2085.685,1375.439,2909.453,2846.163,1508.357,2023.621),nrow=3)

result <- georeference_image(ICP_coordinates, IMAGE_coordinates)
print(result)

}
