% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Calculate and Plot Metrics from a Data Frame}
\usage{
metrics(file, scale = 1, angle_cor = 0)
}
\arguments{
\item{file}{A data frame containing the data to be processed.}

\item{scale}{a scale factor from georeference_image function.}
}
\value{
A ggplot object with histograms of the calculated metrics.
}
\description{
This function calculates additional metrics from a data frame, filters the data, and creates histograms of the metrics.
}
\examples{
df <- data.frame(sample = c('sample_1', 'sample_2'), radius_max = c(10, 20), radius_min = c(2, 5),
                 angle = c(0.5, -0.5), area = c(500, 10000))
plot <- metrics(df)
print(plot)
}
